also use serde in openaddress importer
