+++ importers/openaddresses/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [dependencies]
 csv = "1.1"
+serde = { version = "1.0", features = ["derive"] }
 tools = { path = "../../tools" }
 
 [[bin]]
